extracted (I)ConfigurationSetup in order to make Program smaller and introduced Autofac
