+++ ModernRonin.ProjectRenamer/IConfigurationSetup.cs
@@ -1,0 +1,7 @@
+﻿namespace ModernRonin.ProjectRenamer
+{
+    public interface IConfigurationSetup
+    {
+        (Configuration configuration, string solutionPath) Get(string[] commandLineArguments);
+    }
+}
