+++ ModernRonin.ProjectRenamer/ModernRonin.ProjectRenamer.csproj
@@ -17,6 +17,7 @@
 
 
 	<ItemGroup>
+		<PackageReference Include="Autofac" Version="6.1.0" />
 		<PackageReference Include="ModernRonin.FluentArgumentParser" Version="1.0.0" />
 		<PackageReference Include="morelinq" Version="3.3.2" />
 		<PackageReference Include="Sln" Version="0.3.0" />
